test(api-server): un-quarantine documents Wave G3 residual (BIT-630) (#2542)
* test(api-server): un-quarantine documents Wave G3 residual (BIT-630)

Un-ignore 11 BIT-351 quarantine markers across two suite files; all
target routes exist (reason "schema/route not implemented" was stale).

- documents_intelligence_templates_tests.rs (7): the 6 manager-gated
  template handlers (create/list-via-create/get/update/delete/generate)
  return 403 with a plain login token because TenantExtractor reads the
  singular `role` claim, absent from JwtService login tokens. Switch
  those to create_manager_with_org (mints an OrgAdmin api_core::Claims
  token). submit_classification_feedback has no role gate — un-ignore.
- documents_core_crud_tests.rs (4): update_document, update_document_access,
  create_version, restore_version pass via the creator branch
  (created_by == auth.user_id on the seeded document) — un-ignore.

Zero DELETE (no 404 routes). Ref BIT-354.


* fix(db): cast category/access_scope enums to text in RETURNING clauses (BIT-630)

update_rls and create_version_rls used RETURNING * which returned
category and access_scope as native PG enum types, causing SQLx to
fail decoding into String fields -> 500 INTERNAL_ERROR on update/restore.
Pattern mirrors move_document_rls which already carries the fix.


* fix(api-server): restore organizations/mod.rs formatting to dev baseline (BIT-630)

Local rustfmt version skew had reformatted this file in the previous
commit; restoring the CI-canonical version from origin/dev.


* fix(db): cast enum params in SET clauses and fix placeholder_count bigint (BIT-630)

- update_rls: COALESCE($4::document_category, ...) and ($6::document_access_scope, ...)
- submit_classification_feedback: COALESCE($3::document_category, ...)
- document_template list: jsonb_array_length(...)::bigint for placeholder_count i64 field


---------

+++ backend/crates/db/src/repositories/document_template.rs
@@ -137,7 +137,7 @@ impl DocumentTemplateRepository {
             SELECT
                 id, name, description,
                 template_type::text AS template_type, usage_count,
-                jsonb_array_length(placeholders) as placeholder_count,
+                jsonb_array_length(placeholders)::bigint as placeholder_count,
                 created_at
             FROM document_templates
             WHERE organization_id = $1
